refactor: clean up past_shows folder and update gitignore
