--- packages/frontend/src/pages/past_shows/.ipynb_checkpoints/PastShowsPage-checkpoint.tsx
@@ -1,14 +1,0 @@
-import Card from './components/Card';
-import CardStack from './components/CardStack';
-import './PastShowsPage.css';
-
-export default function PastShowsPage() {
-  return (
-    <>
-      <section className="bg-[#ffc6bf] w-screen h-[100vh]">
-        {/* <Card info={card} /> */}
-      </section>
-      <CardStack />
-    </>
-  );
-}
