--- packages/frontend/src/pages/past_shows/components/.ipynb_checkpoints/Card-checkpoint.css
@@ -1,16 +1,0 @@
-.card {
-  width: 80%;
-  height: 60%;
-  border-radius: 20px;
-  box-shadow: 0px 4px 20px rgba(0, 0, 0, 0.15);
-  position: absolute;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background: white;
-  transition: box-shadow 0.3s ease;
-
-  &:hover {
-    box-shadow: 0px 8px 40px rgba(0, 0, 0, 0.4);
-  }
-}
